Type changed to "double" for being aligned with the constructor of "MetricFamilySamples.Sample"

+++ src/main/java/jp/gr/java_conf/ogibayashi/prometheus/KafkaExporterLogEntry.java
@@ -11,6 +11,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @EqualsAndHashCode(exclude={"value"})
 public class KafkaExporterLogEntry {
     @NonNull private String name;
-    @NonNull private long value;
+    @NonNull private double value;
     private Map<String,String> labels;
 }
